+++ src/app/manager/components/manager-request/manager-request.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core'
-import { ReadManagerRequest } from 'src/app/model/manager-requests/read-manager-request.interface'
+import { ReadManagerRequest } from 'src/app/shared/model/manager-requests/read-manager-request.interface'
 
 @Component({
   selector: 'app-manager-request',
