+++ src/app/shared/components/search-bar/search-bar.component.scss
