+++ src/app/shared/model/events/read-event.interface.ts
@@ -1,4 +1,4 @@
-import { EventStatus } from "../common/event-status.enum"
+import { EventStatus } from "../../shared/common/event-status.enum"
 
 export interface ReadEvent {
     event_id: number
