+++ src/app/shared/model/manager-requests/create-manager-request.interface.ts
