+++ src/app/shared/model/manager-requests/update-manager-request.interface.ts
@@ -1,4 +1,4 @@
-import { RequestStatus } from "../common/request-status.enum"
+import { RequestStatus } from "../../shared/common/request-status.enum"
 
 export interface UpdateManagerRequest {
     status: RequestStatus
