+++ src/app/shared/model/menus/read-menu.interface.ts
