+++ src/app/shared/model/store-requests/read-store-request.interface.ts
