+++ src/app/shared/model/stores/read-store.interface.ts
