+++ src/app/shared/module/common.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core'
 import { CommonModule } from '@angular/common'
 import { FormsModule } from '@angular/forms'
-import { SearchBarComponent } from './components/search-bar/search-bar.component'
-import { FooterMenuComponent } from './components/footer-menu/footer-menu.component'
+import { SearchBarComponent } from '../components/search-bar/search-bar.component'
+import { FooterMenuComponent } from '../components/footer-menu/footer-menu.component'
 import { IonicModule } from '@ionic/angular'
 
 @NgModule({
