+++ src/app/shared/services/manager-requests.service.ts
@@ -2,9 +2,9 @@ import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
 import { ApiResponseDTO } from '../model/common/api-response.interface'
 import { Observable } from 'rxjs'
+import { CreateManagerRequestDTO } from '../model/manager-requests/create-manager-request.interface'
 import { ReadManagerRequest } from '../model/manager-requests/read-manager-request.interface'
 import { UpdateManagerRequest } from '../model/manager-requests/update-manager-request.interface'
-import { CreateManagerRequestDTO } from '../model/manager-requests/create-manager-request.interface'
 
 @Injectable({
   providedIn: 'root'
