+++ src/app/store/components/reviews/reviews.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core'
-import { ReadReview } from 'src/app/model/reviews/read-review.interface'
+import { ReadReview } from 'src/app/shared/model/reviews/read-review.interface'
 
 @Component({
   selector: 'app-reviews',
