+++ src/app/store/components/store-info/store-info.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core'
 import { ActivatedRoute, Router } from '@angular/router'
-import { ReadStore } from 'src/app/model/stores/read-store.interface'
+import { ReadStore } from 'src/app/shared/model/stores/read-store.interface'
 
 @Component({
   selector: 'app-store-info',
