+++ src/app/store/pages/menu-page/menu.page.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core'
 import { ActivatedRoute } from '@angular/router'
-import { ApiResponseDTO } from 'src/app/model/common/api-response.interface'
-import { ReadMenu } from 'src/app/model/menus/read-menu.interface'
-import { ReadStore } from 'src/app/model/stores/read-store.interface'
-import { StoresService } from 'src/app/services/stores.service'
+import { ApiResponseDTO } from 'src/app/shared/model/common/api-response.interface'
+import { ReadMenu } from 'src/app/shared/model/menus/read-menu.interface'
+import { ReadStore } from 'src/app/shared/model/stores/read-store.interface'
+import { StoresService } from 'src/app/shared/services/stores.service'
 
 @Component({
   selector: 'app-menu',
