+++ src/app/store/pages/store-detail-page/store-detail-page.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core'
 import { ActivatedRoute } from '@angular/router'
-import { ApiResponseDTO } from 'src/app/model/common/api-response.interface'
-import { ReadStore } from 'src/app/model/stores/read-store.interface'
-import { StoresService } from 'src/app/services/stores.service'
+import { ApiResponseDTO } from 'src/app/shared/model/common/api-response.interface'
+import { ReadStore } from 'src/app/shared/model/stores/read-store.interface'
+import { StoresService } from 'src/app/shared/services/stores.service'
 
 @Component({
   selector: 'app-store-detail-page',
